third_party/WebKit/LayoutTests/http/tests/inspector/tracing/timeline-xhr-event.html

<html>
<head>
<script src="../inspector-test.js"></script>
<script src="../timeline-test.js"></script>
<script src="../network-test.js"></script>
<script>

function performActions(callback)
{
    var xhr = new XMLHttpRequest();
    xhr.open("GET", "network/resources/resource.php", true);
    xhr.onload = callback;  // This is necessary for XHRLoad event.
    xhr.onreadystatechange = function () { };  // This is necessary for XHRReadyStateChange event.
    xhr.send(null);
}

function test()
{
    InspectorTest.invokeAsyncWithTimeline("performActions", finish);

    function finish()
    {
        InspectorTest.printTimelineRecordsWithDetails("XHRReadyStateChange");
        InspectorTest.printTimelineRecordsWithDetails("XHRLoad");
        InspectorTest.completeTest();
    }
}

</script>
</head>

<body onload="runTest()">
<p>
Tests the Timeline events for XMLHttpReqeust
</p>

</body>
</html>
